Updates to pricing and billing.
Eliminate annual billing options. Add a public pricing page.
Add an internal pricing page. Both are based on the same
table, so updating it once hits both locations.

Place a delay on public product and doc menu to help
with accidental hovering.

Add emails alerts for a newly created subscription, when
a subscription is cancelled, and when a subscription is
deleted after running out.

Add a catcher function to detect when ever the CSRF token
for a given form or page is stale, invalid, incorrect.  If so,
the page is reloaded and the user is warned.